Everything is finally working. Tests are passing (some intermittent crashes due to bad inherited code).
