+++ qiskit/providers/aer/openpulse/cy/CMakeLists.txt
@@ -32,5 +32,5 @@ add_cython_module(memory)
 add_cython_module(utils)
 add_cython_module(numeric_integrator_wrapper src/numeric_integrator.cpp)
 #if(BUILD_TESTS)
-    add_cython_module(test_py_to_cpp_helpers)
+    add_cython_module(test_python_to_cpp)
 #endif()
